use f64 on movement states, rename some messages

+++ src/state/panning.rs
@@ -5,9 +5,9 @@ pub struct State {
     // status which tells if moving is in progress
     pub status: Status,
     // origin of movement
-    origin: (i32, i32),
+    origin: (f64, f64),
     // position of movement
-    position: (i32, i32),
+    position: (f64, f64),
     // velocity of movement
     velocity: (f64, f64),
     // time
@@ -16,7 +16,7 @@ pub struct State {
 
 impl State {
     // begins movement, storing origin point
-    pub fn begin(&mut self, xy: (i32, i32)) {
+    pub fn begin(&mut self, xy: (f64, f64)) {
         self.status = Status::Panning;
         self.origin = xy.clone();
         self.position = xy;
@@ -29,12 +29,12 @@ impl State {
         self.velocity
     }
     // ends state. returns position
-    pub fn end(&mut self) -> (i32, i32) {
+    pub fn end(&mut self) -> (f64, f64) {
         self.status = Status::Idle;
         self.offset()
     }
     // applies movement
-    pub fn set_position(&mut self, xy: (i32, i32)) {
+    pub fn set_position(&mut self, xy: (f64, f64)) {
         self.position = xy;
 
         // buildup velocity
@@ -43,14 +43,14 @@ impl State {
         self.velocity = (o.0 as f64 / dt, o.1 as f64 / dt);
         self.time += dt;
     }
-    pub fn add_relative(&mut self, xy: (i32, i32)) {
+    pub fn add_relative(&mut self, xy: (f64, f64)) {
         self.position.0 += xy.0;
         self.position.1 += xy.1;
     }
     pub fn status(&self) -> Status {
         self.status
     }
-    pub fn offset(&self) -> (i32, i32) {
+    pub fn offset(&self) -> (f64, f64) {
         (
             self.position.0 - self.origin.0,
             self.position.1 - self.origin.1,
